Experimenting with alternative structure to allow for flexible descendant OgnlRuntime processing: - Refactored OgnlRuntime to support alternative processing flow in a more generic way,   providing a convention and method for specialized processing.  Moved the new Method deny processing   out of the base class. - Createed a new OgnlRuntimeMethodBlocking descendant class, which contains the specific Method deny   processing. - Descendants that follow the convention can override the specialized processing method to provide   specific behaviour.  An instance of a (singleton) descendant can be set in the base class, but   this can only be done once. - Updated unit test to correspond to refactoring.
